pyproject.toml
[project]
name = "metalstat"
version = "0.1.0"
description = "Apple Silicon GPU/CPU/Memory monitoring CLI — like gpustat, but for Metal"
readme = "README.md"
authors = [
    { name = "ran", email = "hzz5361@psu.edu" }
]
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["apple-silicon", "gpu", "metal", "monitoring", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
dependencies = [
    "psutil>=5.9.0",
    "blessed>=1.17.1",
]

[project.optional-dependencies]
metal = ["pyobjc-framework-Metal>=10.0"]
all = ["pyobjc-framework-Metal>=10.0"]
dev = ["pytest>=7.0"]

[project.scripts]
metalstat = "metalstat.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
